consolidate db.schemas module
